Fix warning when compiling the sim.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@15858 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/wma.c
@@ -521,7 +521,7 @@ next_track:
                 ci->seek_complete();
                 goto next_track;
             }
-            DEBUGF("Seek returned %d\n", time);
+            DEBUGF("Seek returned %d\n", (int)time);
             ci->set_elapsed(elapsedtime);
 
             /*flush the wma decoder state*/
